Authentication
This JavaScript file manages the authentication for both the login and register pages. It uses Passport.js for validating the forms and MySQL for storing user data. The HTML & CSS are rendered using EJS. 

Flash messages are displayed in both the login and register pages for when the user enters an already existing username (register page) or either the wrong username/password (login page). Successful authentication will redirect the user to a profile page that displays a message which contains their username - I will implement a functioning profile page and a logout option in the next version.