Authentication (using Mongoose ORM)
Just like with the previous v2, this authentication system uses Passport.js, along with other Node.js libraries, to validate forms. The only difference with this one is that it uses MongoDB Atlas as it's database instead of MySQL.

The two JavaScript app files contained within this project use different ways to connect to MongoDB Atlas; In this first app file, the 'mongoose' tool is used, however in the second app file the 'mongodb' driver is used as an alternative - both carry out the same task of inserting form data but they are there to show that I can work with either one.